feat: add dataset download functionality
- Introduced a dedicated dataset download module (utils/raw_data_downloader.py), separated from cleaners (cleaners/basic_cleaners.py)
- Added corresponding usage examples in README
