+++ tidymut/__init__.py
@@ -32,3 +32,10 @@ from .utils.data_source import (
     list_datasets_with_built_in_cleaners,
     show_download_instructions,
 )
+
+from .utils.raw_data_downloader import (
+    download,
+    download_cdna_proteolysis_source_file,
+    download_protein_gym_source_file,
+    download_human_domainome_source_file,
+)
