+++ tidymut/cleaners/cdna_proteolysis_cleaner.py
@@ -41,9 +41,11 @@ class CDNAProteolysisCleanerConfig(BaseCleanerConfig):
 
     Inherits from BaseCleanerConfig and adds cDNAProteolysis-specific configuration options.
 
-    Raw cDNAProteolysis file can be downloaded from
-    - https://zenodo.org/records/7992926, File `Tsuboyama2023_Dataset2_Dataset3_20230416.csv` in `Processed_K50_dG_datasets.zip`
-    - https://huggingface.co/datasets/xulab-research/TidyMut/blob/main/cDNA_proteolysis/Tsuboyama2023_Dataset2_Dataset3_20230416.csv
+    Simply run `tidymut.download_cdna_proteolysis_source_file()` to download the dataset.
+
+    Alternatively, the raw cDNAProteolysis file can be obtained from:
+    - Zenodo: https://zenodo.org/records/7992926, File `Tsuboyama2023_Dataset2_Dataset3_20230416.csv` in `Processed_K50_dG_datasets.zip`
+    - Hugging Face: https://huggingface.co/datasets/xulab-research/TidyMut/blob/main/cDNA_proteolysis/Tsuboyama2023_Dataset2_Dataset3_20230416.csv
 
     Attributes
     ----------
