+++ tidymut/cleaners/human_domainome_cleaner.py
@@ -50,9 +50,11 @@ class HumanDomainomeCleanerConfig(BaseCleanerConfig):
 
     Inherits from BaseCleanerConfig and adds HumanDomainome-specific configuration options.
 
-    Raw HumanDomainome dataset DataFrame or file path to K50 HumanDomainome
-    - File: `SupplementaryTable4.txt` from the article
-        'Site-saturation mutagenesis of 500 human protein domains'
+    Simply run `tidymut.download_human_domainome_source_file()` to download the dataset.
+
+    Alternatively, the raw HumanDomainome file can be obtained from:
+    - Nature artical: 'Site-saturation mutagenesis of 500 human protein domains', File `SupplementaryTable4.txt`
+    - Hugging Face: https://huggingface.co/datasets/xulab-research/TidyMut/blob/main/human_domainome/SupplementaryTable4.txt
 
     Attributes
     ----------
